validation for not deleting the directory

+++ Fishing/get_offer.php
@@ -9,10 +9,8 @@ $br1=-1;
 $now = new DateTime();
 $now->modify('+2 hours');
 $izpishi = $now->format('Y-m-d H:i');
-echo "Now is ".$izpishi."<br>";
 if ($result && mysqli_num_rows($result) > 0) {//look at the OFFERS
     while($row = mysqli_fetch_assoc($result)) {//cycle through ever offer 
-	echo "Ofertata e ".$row["Time"]."<br>";
         if($izpishi  > $row["Time"]){//Search for passed offers
 			$br=$br+1;
 			$delete_id[$br]=$row["ID"];
@@ -40,7 +38,7 @@ foreach($delete_id as $a){
 
 foreach($delete_img as $img){
 	$path="Fishing/Img/Post_Img/".$img;
-	if (file_exists($path)) {
+	if (file_exists($path) && $path !="Fishing/Img/Post_Img/") {
 		unlink($path);
 	}
 }
